Create interface.java
Java Does not support Multiple Inheritance so we instead of using Interface. Interface in a blueprint of a class. It has static constants and Abstract method. keyword--> interface
It provides total Abstraction.
it means all the methods in an interface are declared with body and all the fields are public, static and final by default.
A class implements an interface, but one interface extends another interface.